refactor: 컴포넌트 및 messages 분리 - #3
- 지원 과제 안내 Box 컴포넌트 분리
- 메인페이지에서 사용되는 파트별 메세지 분리
