+++ src/messages/part.js
@@ -1,0 +1,20 @@
+const part = {
+  web: {
+    part_name: 'Web',
+    content:
+      '실제 서비스에 사용되는 API 서버를 설계하고, NestJS를 통해 실제 작동하는 서버로 구현/배포해요.',
+    tech_list: 'git, Github, JavaScript, Netlify, React',
+    assignment:
+      'HTML, CSS를 사용해 지원자를 소개하는 자기소개 페이지를 만들어 제출해주세요.',
+  },
+  server: {
+    part_name: 'Server',
+    content:
+      '실제 서비스에 사용되는 API 서버를 설계하고, NestJS를 통해 실제 작동하는 서버로 구현/배포해요.',
+    tech_list:
+      'git, Github, Spring Boot, JPA, Spring Security, JAVA, nginx, EC2, MySQL',
+    assignment: '간단한 Java 프로그래밍 과제를 제출해주세요.',
+  },
+};
+
+export default part;
